feat: fix haskell and include rust
