+++ internal/sandbox/templates/haskell.txt
@@ -1,2 +1,1 @@
-main = do
-  print "Hello, World!"
+main = putStrLn "Hello, World!"
