+++ internal/sandbox/templates/tests/multi/haskell.txt
@@ -1,0 +1,6 @@
+hello = "Hello"
+world = "World"
+
+main = putStrLn (hello ++ ", " ++ world ++ "!")
+
+
